Put subQueue selection into an interface so that we can choose a different implementation.
LBMQ chooses the next queue based on priority queues. It chooses the queue from the highest priority priority group. If not found it searches the lower priority groups until it finds it.
This behavior does not always suit customer needs. Therefore, I introduced and interface which allows us to provide a different queue selection algorithm.